# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerIn": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AnswerIn",
        "type": "object"
      },
      "ApplyIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ApplyIn",
        "type": "object"
      },
...
      "CredentialIn": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "model": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "default": "openai",
            "maxLength": 32,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "CredentialIn",
        "type": "object"
      },
      "CredentialOut": {
        "properties": {
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "base_url",
          "model",
          "configured"
        ],
        "title": "CredentialOut",
        "type": "object"
      },
...
      "DraftIn": {
        "additionalProperties": false,
        "properties": {
          "draft": {
            "additionalProperties": true,
            "title": "Draft",
            "type": "object"
          }
        },
        "required": [
          "draft"
        ],
        "title": "DraftIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartIn": {
        "additionalProperties": false,
        "properties": {
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "title": "StartIn",
        "type": "object"
      },
      "SummaryIn": {
        "additionalProperties": false,
        "properties": {
          "theses": {
            "items": {
              "type": "string"
            },
            "title": "Theses",
            "type": "array"
          }
        },
        "required": [
          "theses"
        ],
        "title": "SummaryIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Planora",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ai/credential": {
      "get": {
        "operationId": "read_credential_api_ai_credential_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Credential",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "write_credential_api_ai_credential_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Credential",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions": {
      "post": {
        "operationId": "start_session_api_ai_sessions_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_api_ai_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions/{session_id}/answers": {
      "post": {
        "operationId": "answer_question_api_ai_sessions__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Question",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions/{session_id}/apply": {
      "post": {
        "description": "Step 4: applying as a batch of mutations with a shared batch_id.",
        "operationId": "apply_session_api_ai_sessions__session_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Session",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions/{session_id}/draft": {
      "post": {
        "description": "Gate 2, the main one: the draft. Nothing has been written into the project.",
        "operationId": "build_draft_api_ai_sessions__session_id__draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Draft",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "edit_draft_api_ai_sessions__session_id__draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Draft",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/sessions/{session_id}/summary": {
      "post": {
        "description": "Gate 1: \"here is what I understood about the project\".",
        "operationId": "build_summary_api_ai_sessions__session_id__summary_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Summary",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "edit_summary_api_ai_sessions__session_id__summary_put",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Summary",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/tasks/{task_id}/split": {
      "post": {
        "description": "A suggestion to split a task. Writes nothing: it is applied by a button.",
        "operationId": "propose_split_api_ai_tasks__task_id__split_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Split",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/tasks/{task_id}/split/apply": {
      "post": {
        "operationId": "apply_split_api_ai_tasks__task_id__split_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "planora_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planora Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Split",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}